WIP - after splitting contour.h.cxx into more manageable pieces.
But still failing test in testContour.cxx::ClosePolygon)...
